server: add skeleton for server logging
An idea is to pass log inside context to all underlying functions.
Log context is initialized at the top-most application level.
Common context package is separated to allow access to it from other packages.
Third-party packages which do not support slog will be directed to it via a configured bridge.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>