WIP: gateway: add docker creds handler.
I'm not sure if we will need it, so no unit tests.
The reason I took this was promarily to play with opaque tokens which NEED NOT be stored in the database.
For that, we can use simple AES-GSM encrypted tokens bearing all the needed info.
We may consider a JWT-E instead, but I did not find a good Golang library with quick search.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>