api: save and return effective device UUIDs participating in the rollout
When a rollout happens, there are two specifics requiring attention.:
- Not all device UUIDs specified in the rollout are updated - only those matching the update tag are.
- If device group participation changes after the rollout - device update name is not changed.

It means that we need a sort of a capture for device UUIDs that were effectively updated for a given rollout.
That allows a user to know for sure which devices were scheduled for an update, and which were not.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>