storage: write files to disk safely
A safe write is performed via a write into a temporary file followed by a file move.
A suffix for a temporary file is selected to be consistent with modern browsers upload suffixes.
There is an intentional double-dot in the suffix to make collisions with user-named files improbable.

Such files are filtered out during the file listing.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>