api: spawn rollout database update asynchronously
Updating devices can take some time, so run it async.
There is no harm if it fails, as the background daemon will replay all failed ops.

We may consider a worker thread pool in the future, but for now spanwing a new goroutine is just fine.
A legitimate user is not expected to spam its own API server with quadzillion rollout requests to DoS it.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>