web: Convert to HTML templates
This uses the very minimal picoss framework that's based on semantic
HTML. It's small enough that we can move to something else later on.

We will eventually have to add the CSS file locally so this can work
in a true offline fashion. However, for the current development phase,
I'm adding it as a link to keep things easier.

Signed-off-by: Andy Doan <andy@foundries.io>