log: fixup: properly set the log level
This is actually a bug fix.
We were setting a level at which standard logger should log messages.
And we were always using the default level to filter which messages should be logged.

This commit correctly sets the log level to filter messages.

It also sets the standard logger messages to always be at warning level.
Eventually, we'd like those to be cleaned up somehow (if they ever appear in our logs).

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>