mdns: auto-detect the server IP address from available interfaces
A user can still provide the API address for the gateway manually.
But, they can also provide the interface name, so that the IP address is auto-detected.
If both IPv4 and IPv6 address are configured for the interface, the latter takes precedence.
If a user specified neither address nor interface, auto-detect a non-loopback interface to use.
If there is no non-loopback interfce or more than one non-loopback interface - require to provide either interface or address.
Note that a user can still specify the loopback interface manually.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>